fix(factor_analysis): LEAD horizon 参数改 i32 修复实跑报错
实跑发现:LEAD(close, horizon) 第二参数 PostgreSQL 要求 integer,
原绑定 i64(bigint)报 'function lead(numeric, bigint) does not exist'。改 i32 修复。

实跑验证:fin_roe_daily_std [1,5,10,20,60] 衰减曲线跑通,IC 随 horizon 单调变负。

+++ quant-api/src/routes/factor_analysis.rs
@@ -150,7 +150,7 @@ pub async fn fetch_forward_returns(
     .bind(symbols)
     .bind(start)
     .bind(end)
-    .bind(horizon as i64)
+    .bind(horizon as i32)
     .fetch_all(db)
     .await
     .map_err(|e| format!("fetch forward_returns(h={}): {}", horizon, e))?;
